update ProductPage and ProfilePage
